add fix

+++ Guild/ApplyGuild.js
@@ -26,7 +26,7 @@ var time
 // }
 exports.S_APPLY_GUILD = function s_apply_guild (socket,data) {
 	checkCurrentGuild (socket,data,function (checkGuildBool) {
-		if (checkGuildBool==false) {
+		if (checkGuildBool==true) {
 			//console.log('chua co guild');
 			applyGuild (socket,data);
 		}
@@ -40,6 +40,7 @@ function applyGuild (socket,data) {
 		data.Might  = rows[0].Might;
 		data.Killed = rows[0].Killed;
 		data.NameInGame = rows[0].NameInGame;
+		data.AcceptTime = timeAccept;
 		//console.log(data)
 		var stringInsertApply = "INSERT INTO `"+data.Guild_ID+"`(`ID_User`, `NameInGame`, `Might`, `Killed`, `AcceptTime`) VALUES ('"
 		+data.ID_User+"','"
@@ -49,7 +50,7 @@ function applyGuild (socket,data) {
 		+functions.ImportTimeToDatabase(new Date(timeOutApply).toISOString())+"')";
 
 		db_all_guild.query(stringInsertApply,function (error,result) {
-			if (!!error) {console.log(error)};
+			if (!!error){DetailError = ('ApplyGuild.js: applyGuild: '+ stringInsertApply);functions.WriteLogError(DetailError,2);}
 			sendApplyToGuildMember (socket,data);
 		});
 		setTimeAccept(timeAccept,data,1);		
@@ -59,17 +60,19 @@ function sendApplyToGuildMember (socket,data) {
 	var stringQuery = "SELECT `ID_User` FROM `"+data.Guild_ID+"`";
 	// console.log(stringQuery);
 	db_all_guild.query(stringQuery,function (error,rows) {
+		if (!!error){DetailError = ('ApplyGuild.js: sendApplyToGuildMember: '+ stringQuery);functions.WriteLogError(DetailError,2);}
 		if (rows.length>0) {
 			for (var i = 0; i < rows.length; i++) {
-				getUserInfo (socket,rows[i],data);
+				R_APPLY (socket,rows[i],data);
 			}
 		}
 	});
 }
-function getUserInfo (socket,row,data) {
+function R_APPLY (socket,row,data) {
 	// console.log(data.ID_User)
-	var stringQuery = "SELECT `Server_ID`,`Socket` FROM `user_info` WHERE `ID_User`='"+row.ID_User+"'";
+	var stringQuery = "SELECT `Socket` FROM `user_info` WHERE `ID_User`='"+row.ID_User+"'";
 	db_all_user.query(stringQuery,function (error,rows) {
+		if (!!error){DetailError = ('ApplyGuild.js: sendApplyToGuildMember: '+ stringQuery);functions.WriteLogError(DetailError,2);}
 		if (rows[0].Socket!=null) {
 			socket.broadcast.to(rows[0].Socket).emit('R_APPLY',{R_APPLY:data});
 		}
@@ -79,8 +82,8 @@ function setTimeAccept (timeOutApply,data,enumTime) {
 	var stringTimeOut =  data.ID_User+"_"+data.Guild_ID+"_"+enumTime;
 	DictTimeOut[stringTimeOut] = setTimeout(function (data){
 		var stringCheckGuild = "SELECT * FROM `"+data.Guild_ID+"` WHERE `ID_User`='"+data.ID_User+"'";
-		
 		db_all_guild.query(stringCheckGuild, function (error,rows) {
+			if (!!error){DetailError = ('ApplyGuild.js: setTimeAccept: '+ stringCheckGuild);functions.WriteLogError(DetailError,2);}
 			if (rows.length>0) {
 				switch (enumTime) {
 					case 1:
@@ -98,8 +101,7 @@ function setTimeAccept (timeOutApply,data,enumTime) {
 function updateAcceptTime (data,row,enumTime) {
 	if (row.AcceptTime!=null) {
 		var currentTime 	= functions.GetTime();
-		var databaseTime 	= functions.ExportTimeDatabase(row.AcceptTime);
-		
+		var databaseTime 	= functions.ExportTimeDatabase(row.AcceptTime);		
 		if (databaseTime>currentTime) {
 			var timeOut = databaseTime - currentTime;
 			setTimeAccept (timeOut,data,enumTime);
@@ -109,10 +111,20 @@ function updateAcceptTime (data,row,enumTime) {
 	}
 }
 function deleteApply (Guild_ID,ID_User) {
-	var stringDelete = "DELETE FROM `"+Guild_ID+"` WHERE `ID_User`='"+ID_User+"'";
-	db_all_guild.query(stringDelete, function (error,result) {
-		if (!!error) {console.log(error)};
+	var stringSelect = "SELECT `RemoveTime` FROM `"+Guild_ID+"` WHERE `ID_User`='"+ID_User+"'";
+	var stringDelete;
+	db_all_guild.query(stringSelect,function (error,rows) {
+		if (rows[0].RemoveTime!=null) {
+			stringDelete ="UPDATE `13` SET `AcceptTime` = NULL WHERE `ID_User`='"+ID_User+"'";
+		}else{
+			stringDelete = "DELETE FROM `"+Guild_ID+"` WHERE `ID_User`='"+ID_User+"'";		
+		}
+		db_all_guild.query(stringDelete, function (error,result) {
+			if (!!error){DetailError = ('ApplyGuild.js: deleteApply: '+ stringDelete);functions.WriteLogError(DetailError,2);}
+			LogChange = 'ApplyGuild.js: deleteApply: '+stringDelete;functions.LogChange(LogChange,2);
+		});
 	});
+	
 }
 function updateRemoveTime (data,row,enumTime) {
 	if (row.RemoveTime!=null) {
@@ -125,13 +137,15 @@ function updateRemoveTime (data,row,enumTime) {
 		}else {
 			var stringDelete = "DELETE FROM `"+data.Guild_ID+"` WHERE `ID_User`='"+data.ID_User+"'";
 			db_all_guild.query(stringDelete, function (error,result) {
-				if (!!error) {console.log(error)};
+				if (!!error){DetailError = ('ApplyGuild.js: updateRemoveTime: '+ stringDelete);functions.WriteLogError(DetailError,2);}
+				LogChange = 'ApplyGuild.js: updateRemoveTime: '+stringDelete;functions.LogChange(LogChange,2);
 			});
 
 			getUserServerID (data.ID_User,function (returnServer){
 				var stringUpdateUserInfo = "UPDATE `game_info_s"+returnServer+"` SET `LastGuildID`=null WHERE `ID_User`='"+data.ID_User+"' AND `LastGuildID`='"+data.Guild_ID+"'";
 				db_all_user.query(stringUpdateUserInfo,function (error,result) {
-					if (!!error) {console.log(error)};
+					if (!!error){DetailError = ('ApplyGuild.js: stringUpdateUserInfo: '+ stringUpdateUserInfo);functions.WriteLogError(DetailError,2);}
+					LogChange = 'ApplyGuild.js: stringUpdateUserInfo: '+stringUpdateUserInfo;functions.LogChange(LogChange,2);
 				});
 			});
 		}
@@ -140,11 +154,12 @@ function updateRemoveTime (data,row,enumTime) {
 function getUserServerID (dataID_User,returnServer) {
 	var stringQueryUser = "SELECT `Server_ID` FROM `user_info` WHERE `ID_User`='"+dataID_User+"'";
 	db_all_user.query(stringQueryUser, function (error,rows) {
-		if (!!error) {console.log(error)};
+		if (!!error){DetailError = ('ApplyGuild.js: getUserServerID: '+ stringQueryUser);functions.WriteLogError(DetailError,2);}
 		returnServer(rows[0].Server_ID)
 	});
 }
-function checkCurrentGuild (socket,data,returnBool) {
+
+function checkCurrentGuild (socket,data,checkGuildBool) {
 	var returnBool = false;
 	var currentGuildBool = false;
 	var guildApplyBool = false;
@@ -152,6 +167,7 @@ function checkCurrentGuild (socket,data,returnBool) {
 	new Promise((resolve,reject)=>{
 		var stringQuery = "SELECT `Guild_ID` FROM `game_info_s"+data.Server_ID+"` WHERE `ID_User`='"+data.ID_User+"'";
 		db_all_user.query(stringQuery,function (error,rows) {
+			if (!!error){DetailError = ('ApplyGuild.js: checkCurrentGuild_stringQuery: '+ stringQuery);functions.WriteLogError(DetailError,2);}
 			if (rows[0].Guild_ID!=null) {
 				currentGuildBool = true;
 				EnumApplyGuild.Enum = 2;
@@ -165,6 +181,7 @@ function checkCurrentGuild (socket,data,returnBool) {
 	new Promise((resolve,reject)=>{		
 		var stringGuildApply = "SELECT `ID_User` FROM `all_guilds`.`"+data.Guild_ID+"` WHERE `ID_User`='"+data.ID_User+"'";
 		db_all_guild.query(stringGuildApply,function (error,rows) {	
+			if (!!error){DetailError = ('ApplyGuild.js: checkCurrentGuild_stringGuildApply: '+ stringGuildApply);functions.WriteLogError(DetailError,2);}
 			if(rows.length>0){
 				guildApplyBool=true;
 				EnumApplyGuild.Enum = 3;
@@ -175,12 +192,12 @@ function checkCurrentGuild (socket,data,returnBool) {
 		});
 
 	}).then(()=>{
-		if (currentGuildBool||guildApplyBool) {
+		if (currentGuildBool==false&&guildApplyBool==false) {
 			returnBool=true; 
 			EnumApplyGuild.Enum = 1;
 			EnumApplyGuild.Message = "Apply Guild Success";
 		}
-		socket.emit('R_APPLY_GUILD',{S_APPLY_GUILD:EnumApplyGuild});
+		socket.emit('R_APPLY_GUILD',{R_APPLY_GUILD:EnumApplyGuild});
 		checkGuildBool(returnBool);
 	}));
 }
@@ -204,6 +221,7 @@ function updateApplyPlayer (data) {
 	new Promise((resolve,reject)=>{
 		var getGuildName = "SELECT `GuildName` FROM `guild_info` WHERE `Guild_ID`='"+data.Guild_ID+"'";
 		db_all_guild.query(getGuildName,function (error,rows) {
+			if (!!error){DetailError = ('ApplyGuild.js: updateApplyPlayer: '+ getGuildName);functions.WriteLogError(DetailError,2);}
 			data.GuildName = rows[0].GuildName;
 			resolve();
 		});
@@ -211,6 +229,8 @@ function updateApplyPlayer (data) {
 		getUserServerID(data.ID_Apply,function (returnServer){
 			var stringUpdateUser = "UPDATE `game_info_s'"+returnServer+"'` SET `Guild_ID`='"+data.Guild_ID+"',`Guild_Name`='"+data.GuildName+"',`LastGuildID`=null";
 			db_all_user.query(stringUpdateUser,function (error,result) {
+				if (!!error){DetailError = ('ApplyGuild.js: stringUpdateUser: '+ stringUpdateUser);functions.WriteLogError(DetailError,2);}
+				LogChange = 'ApplyGuild.js: stringUpdateUser: '+stringUpdateUser;functions.LogChange(LogChange,2);
 				if (!!error) {console.log(error)};
 			});
 		});
@@ -221,9 +241,8 @@ function updateApplyPlayer (data) {
 }
 function sendUserInfo (socket,data) {
 	var stringQueryGuildMember = "SELECT `ID_User` FROM `"+data.Guild_ID+"` WHERE `ID_User`<>'"+data.ID_Player+"'";
-	console.log(stringQueryGuildMember)
 	db_all_guild.query(stringQueryGuildMember, function (error,rows) {
-		console.log(rows);
+		if (!!error){DetailError = ('ApplyGuild.js: stringQueryGuildMember: '+ stringQueryGuildMember);functions.WriteLogError(DetailError,2);}
 		for (var i = 0; i < rows.length; i++) {
 			R_ACCEPT_APPLY (socket,rows[i].ID_User);
 		}
@@ -232,6 +251,7 @@ function sendUserInfo (socket,data) {
 function R_ACCEPT_APPLY (socket,data) {
 	var stringQuery = "SELECT `Socket` FROM `user_info` WHERE `ID_User`='"+data+"'";
 	db_all_user.query(stringQuery,function (error,rows) {
+		if (!!error){DetailError = ('ApplyGuild.js: R_ACCEPT_APPLY: '+ stringQuery);functions.WriteLogError(DetailError,2);}
 		for (var i = 0; i < rows.length; i++) {
 			socket.broadcast.to(rows[i].Socket).emit('R_ACCEPT_APPLY',{R_ACCEPT_APPLY:data});
 		}
@@ -242,7 +262,7 @@ function updateGuildAccept (data) {
 		getUserServerID (data.ID_Apply,function (returnServer) {
 			var stringUserMight ="SELECT `Might` FROM `game_info_s"+returnServer+"` WHERE `ID_User`='"+data.ID_Apply+"'";
 			db_all_user.query(stringUserMight,function (error, rows) {
-				if (!!error) {console.log(error)};
+				if (!!error){DetailError = ('ApplyGuild.js: updateGuildAccept_stringUserMight: '+ stringUserMight);functions.WriteLogError(DetailError,2);}
 				data.Might = rows[0].Might;
 				resolve();
 			});
@@ -250,14 +270,14 @@ function updateGuildAccept (data) {
 	}).then(()=>{
 		var updateGuild = "UPDATE `"+data.Guild_ID+"` SET `AcceptTime`= null,`RemoveTime`= null WHERE `ID_User`='"+data.ID_Apply+"';"
 		+"UPDATE `guild_info` SET `Member`=`Member`+1,`Might`=`Might`+'"+data.Might+"' WHERE `Guild_ID`='"+data.Guild_ID+"';";
-		console.log(updateGuild)
 		var stringTimeOut =  data.ID_Apply+"_"+data.Guild_ID+"_1";
 		if (stringTimeOut in DictTimeOut){
 			clearTimeout(stringTimeOut);
 			delete DictTimeOut[stringTimeOut];
 		}
 		db_all_guild.query(updateGuild,function (error,result) {
-			if (!!error) {console.log(error)};
+			if (!!error){DetailError = ('ApplyGuild.js: updateGuildAccept_updateGuild: '+ updateGuild);functions.WriteLogError(DetailError,2);}
+			LogChange = 'ApplyGuild.js: updateGuildAccept_updateGuild: '+updateGuild;functions.LogChange(LogChange,2);
 		});
 	});
 }
@@ -277,11 +297,22 @@ function checkGuildPosition (data,checkBool) {
 		getUserServerID (data.ID_Apply,function (returnServer) {
 			var stringPlayerGuild = "SELECT `Guild_ID` FROM `game_info_s"+returnServer+"` WHERE `ID_User`='"+data.ID_Apply+"'";
 			db_all_user.query(stringPlayerGuild,function (error,rows) {
+				if (!!error){DetailError = ('ApplyGuild.js: stringPlayerGuild: '+ stringPlayerGuild);functions.WriteLogError(DetailError,2);}
 				if (rows_stringPlayerGuild[0].Guild_ID==null) {
 					applyGuildBool=true;
 				}else{
-					console.log('da co guild => xoa, clear timeout');
+					//console.log('da co guild => xoa, clear timeout');
 					deleteApply(data.Guild_ID,data.ID_Apply);
+					var stringTimeOut =  data.ID_User+"_"+data.Guild_ID+"_1";
+					if (DictTimeOut[stringTimeOut]!=undefined) {
+						clearTimeout(stringTimeOut);
+						delete DictTimeOut[stringTimeOut];
+					}
+					var dataRej ={
+						ID_Reject: data.ID_Apply,
+						Guild_ID: data.Guild_ID
+					}
+					sendReject (socket,dataRej);
 				}
 				resolve();
 			});
@@ -295,9 +326,58 @@ function getGuildPosition (Guild_ID,ID_User,returnPosition) {
 	var position = 1;
 	var stringQuery = "SELECT `GuildPosition` FROM `"+Guild_ID+"` WHERE `ID_User`='"+ID_User+"'";
 	db_all_guild.query(stringQuery,function (error,rows) {
-		if (!!error) {console.log(error)};
+		if (!!error){DetailError = ('ApplyGuild.js: getGuildPosition: '+ stringQuery);functions.WriteLogError(DetailError,2);}
 		position = rows[0].GuildPosition;
 		returnPosition(position);
 	});
 }
 
+var dataReject ={
+	ID_User: 9,
+	ID_Reject: 42,
+	Guild_ID: 13
+}
+exports.S_REJECT_APPLY = function s_reject_apply(socket,data) {
+	var positionBool = false;
+	new Promise((resolve,reject)=>{
+		getGuildPosition (data.Guild_ID,data.ID_User,function (returnPosition) {
+			if (returnPosition>3) {
+				positionBool=true;
+				deleteApply(data.Guild_ID,data.ID_Reject);
+				var stringTimeOut =  data.ID_User+"_"+data.Guild_ID+"_1";
+				if (DictTimeOut[stringTimeOut]!=undefined) {
+					clearTimeout(stringTimeOut);
+					delete DictTimeOut[stringTimeOut];
+				}
+			}
+			resolve();
+		});
+	}).then(()=>{
+		sendReject (socket,data);
+	});
+}
+
+function sendReject (socket,data) {
+	var stringQuery = "SELECT `ID_User` FROM `"+data.Guild_ID+"`";
+	db_all_user.query(stringQuery,function (error,rows) {
+		if (rows.length>0) {
+			for (var i = 0; i < rows.length; i++) {				
+				R_REJECT_APPLY (socket,data,rows[i])
+			}
+		}
+	});
+}
+
+function R_REJECT_APPLY (socket,data,row) {
+	var stringQuery = "SELECT `Socket` FROM `user_info` WHERE `ID_User`='"+row.ID_User+"'";
+	var dataRe ={
+		ID_Reject: data.ID_Reject,
+		Guild_ID: data.Guild_ID
+	}
+	db_all_user.query(stringQuery,function (error,rows) {
+		if (!!error){DetailError = ('ApplyGuild.js: sendApplyToGuildMember: '+ stringQuery);functions.WriteLogError(DetailError,2);}
+		if (rows[0].Socket!=null) {
+			socket.broadcast.to(rows[0].Socket).emit('R_REJECT_APPLY',{R_REJECT_APPLY:dataRe});
+		}
+	});
+}
